+++ app/src/main/java/com/example/meal_prep_planner_app/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.annotation.RequiresApi
+import com.example.meal_prep_planner_app.ui.navigation.AppNavigation
 import com.example.meal_prep_planner_app.ui.theme.MealPrepPlannerAppTheme
 import dagger.hilt.android.AndroidEntryPoint
