+++ app/src/main/java/com/example/meal_prep_planner_app/ui/navigation/AppNavigation.kt
@@ -1,4 +1,4 @@
-package com.example.meal_prep_planner_app
+package com.example.meal_prep_planner_app.ui.navigation
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.meal_prep_planner_app.screens.LoginScreen
+import com.example.meal_prep_planner_app.ui.screens.LoginScreen
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
