+++ app/src/main/java/com/example/meal_prep_planner_app/ui/screens/ProfileScreen.kt
@@ -1,4 +1,4 @@
-package com.example.meal_prep_planner_app.screens
+package com.example.meal_prep_planner_app.ui.screens
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
